+++ src/document/renderer.ml
@@ -42,3 +42,7 @@ let document_of_compilation_unit ~syntax v =
   match syntax with
   | Reason -> Reason.compilation_unit v
   | OCaml -> ML.compilation_unit v
+
+let document_of_asset path (v : Odoc_model.Lang.Asset.t) =
+  let url = Url.Path.from_identifier v.name in
+  Types.Document.Asset { url; src = path }